Time: 11 ms (67.04%), Space: 50 MB (88.53%) - LeetHub

+++ 1011-capacity-to-ship-packages-within-d-days/1011-capacity-to-ship-packages-within-d-days.java
@@ -19,7 +19,7 @@ class Solution {
     }
 
     public boolean isPossible(int[] weights, int capacity, int limit) {
-        long total = 0;
+        int total = 0;
         int days = 1;
         for (int weight : weights) {
             total += weight;
